Extracted ActionLog to its own package.
