+++ kansha/notifications.py
@@ -10,20 +10,13 @@
 
 from __future__ import absolute_import
 
-from datetime import datetime, timedelta
-import json
 import urlparse
-
-from elixir import using_options
-from elixir import ManyToOne
-from elixir import Field, Unicode, UnicodeText, DateTime
-from nagare import database, log, presentation, var, ajax
-from nagare.i18n import _, _L, format_datetime
-from nagare.namespaces import xhtml
-from sqlalchemy.types import TypeDecorator
 import sqlalchemy as sa
 
-from kansha.models import Entity
+from nagare import database
+from nagare.i18n import _, _L
+from nagare.namespaces import xhtml
+
 from kansha.user.models import DataBoardMember
 
 
@@ -33,24 +26,7 @@ NOTIFY_MINE = 1
 NOTIFY_ALL = 2
 
 
-# Groups and messages
-EVENT_MESSAGES = {
-    'card_create': _L(u'Card "%(card)s" has been added to column "%(column)s"'),
-    'card_delete': _L(u'User %(author)s has deleted card "%(card)s"'),
-    'card_archive': _L(u'User %(author)s has archived card "%(card)s"'),
-    'card_move': _L(u'Card "%(card)s" has been moved from column "%(from)s" to column "%(to)s"'),
-    'card_title': _L(u'Card "%(from)s" has been renamed to "%(to)s"'),
-    'card_weight': _L(u'Card "%(card)s" has been weighted from (%(from)s) to (%(to)s)'),
-    'card_add_member': _L(u'User %(user)s has been assigned to card "%(card)s"'),
-    'card_remove_member': _L(u'User %(user)s has been unassigned from card "%(card)s"'),
-    'card_add_comment': _L(u'User %(author)s has commented card "%(card)s"'),
-    'card_add_file': _L(u'User %(author)s has added file "%(file)s" to card "%(card)s"'),
-    'card_add_list': _L(u'User %(author)s has added the checklist "%(list)s" to card "%(card)s"'),
-    'card_delete_list': _L(u'User %(author)s has deleted the checklist "%(list)s" from card "%(card)s"'),
-    'card_listitem_done': _L(u'User %(author)s has checked the item %(item)s from the checklist "%(list)s", on card "%(card)s"'),
-    'card_listitem_undone': _L(u'User %(author)s has unchecked the item %(item)s from the checklist "%(list)s", on card "%(card)s"'),
-}
-
+# Groups
 GROUP_EVENTS = (
     ('affected', ('card_add_member', 'card_remove_member')),
     ('modified', ('card_move', 'card_title', 'card_add_file', 'card_add_comment', 'card_weight',
@@ -63,43 +39,6 @@ GROUP_MESSAGES = {
     'modified': _L(u'Cards modification:'),
     'add_remove': _L(u'Cards addition / removal:')
 }
-
-
-# Models
-
-class JSONType(TypeDecorator):
-    impl = UnicodeText
-
-    def process_bind_param(self, value, dialect):
-        if value is not None:
-            value = unicode(json.dumps(value))
-        return value
-
-    def process_result_value(self, value, dialect):
-        if value is not None:
-            value = json.loads(value)
-        return value
-
-
-class DataHistory(Entity):
-    using_options(tablename='history', order_by='-when')
-
-    when = Field(DateTime)
-    action = Field(Unicode(255))
-    data = Field(JSONType)
-
-    board = ManyToOne('DataBoard', ondelete='cascade')
-    card = ManyToOne('DataCard', ondelete='cascade')
-    user = ManyToOne('DataUser', ondelete='cascade')
-
-
-def add_history(board, card, user, action, data):
-    data.update(action=action)
-    when = datetime.utcnow()
-    data = DataHistory(
-        when=when, action=action, board=board, card=card, user=user, data=data)
-    database.session.add(data)
-    database.session.flush()
 
 
 def get_board_member(user, board):
@@ -124,35 +63,6 @@ def get_subscribers():
     return q
 
 
-def get_last_activity(board):
-    q = database.session.query(DataHistory.when)
-    q = q.filter(DataHistory.board == board)
-    q = q.order_by(DataHistory.when.desc())
-    q = q.limit(1)
-    return q.scalar()
-
-
-def get_events(board, hours=None):
-    '''board to None means "everything".'''
-    since = datetime.utcnow() - timedelta(hours=hours)
-    q = DataHistory.query
-    if board:
-        q = q.filter_by(board=board)
-    q = q.filter(DataHistory.when >= since)
-    q = q.order_by(DataHistory.board_id, DataHistory.action, DataHistory.when)
-    return q.all()
-
-
-def get_history(board, cardid=None, username=None):
-    q = DataHistory.query
-    q = q.filter_by(board=board)
-    if cardid:
-        q = q.filter(DataHistory.card.has(id=cardid))
-    if username:
-        q = q.filter(DataHistory.user.has(username=username))
-    return q
-
-
 def filter_events(events, subscriber):
     if subscriber.notify == NOTIFY_ALL:
         return [event for event in events]
@@ -161,28 +71,8 @@ def filter_events(events, subscriber):
         return [event for event in events if event.card_id in user_cards]
     return []
 
-# component
-
-
-class ActionLog(object):
-
-    def __init__(self, board):
-        self.board = board
-        self.user_id = var.Var('')
-        self.card_id = var.Var(None)
-
 
 # renders
-
-def render_event(event):
-    data = event.data.copy()
-    data['author'] = event.user.fullname or event.user.username
-    msg = EVENT_MESSAGES.get(event.action)
-    if msg is not None:
-        return _(msg) % data
-    log.error('Undefined event type "%s"', event.action)
-    return u''
-
 
 def generate_email(app_title, board, user, hours, url, events):
     ret = []
@@ -221,16 +111,16 @@ def generate_email(app_title, board, user, hours, url, events):
                             # IDs are interpreted as anchors since HTML4. So don't use the ID of
                             # the card as a URL fragment, because the browser
                             # jumps to it.
-                            event = h.a(render_event(event), href='%s#id_card_%s' % (
+                            event = h.a(event.to_string(), href='%s#id_card_%s' % (
                                 data['url'], event.card.id), style='text-decoration: none;')
                         else:
-                            event = render_event(event)
+                            event = event.to_string()
                         root.append(h.li(event))
 
             ret.append(_(GROUP_MESSAGES[group]))
             ret.append('')
             for event in events:
-                ret.append(u'- ' + render_event(event))
+                ret.append(u'- ' + event.to_string())
             ret.append(u'')
 
     ret.append(
@@ -240,33 +130,3 @@ def generate_email(app_title, board, user, hours, url, events):
                     _(' and modify notifications parameters in the "settings" menu of the "Board" tab.') % data, style='border: 1px solid black; padding: 5px; text-align: center;'))
 
     return subject, u'\n'.join(ret), root.write_htmlstring()
-
-
-@presentation.render_for(ActionLog)
-def render(self, h, *_args):
-    return h.root
-
-
-@presentation.render_for(ActionLog, 'history')
-def render_history(self, h, *_args):
-    h << h.h2(_('Action log'))
-    board = self.board.data
-    with h.select(onchange=ajax.Update(action=self.user_id)):
-        h << h.option(_('all users'), value='')
-        for member in board.members:
-            h << h.option(member.fullname, value=member.username).selected(
-                self.user_id())
-    with h.select(onchange=ajax.Update(action=lambda x: self.card_id(int(x)))):
-        h << h.option(_('all cards'), value=0)
-        for col in board.columns:
-            for card in col.cards:
-                h << h.option(card.title, value=card.id).selected(
-                    self.card_id())
-    with h.div(class_='history'):
-        with h.table(class_='table table-striped table-hover'):
-            with h.body:
-                for event in get_history(board, cardid=self.card_id(), username=self.user_id()):
-                    with h.tr:
-                        h << h.th(format_datetime(event.when, 'short'))
-                        h << h.td(render_event(event))
-    return h.root
