+++ kansha/services/actionlog/__init__.py
@@ -1,12 +1,12 @@
-#--
+# -*- coding:utf-8 -*-
+# --
 # Copyright (c) 2012-2014 Net-ng.
 # All rights reserved.
 #
 # This software is licensed under the BSD License, as described in
 # the file LICENSE.txt, which you should have received as part of
 # this distribution.
-#--
+# --
 
-from .services_repository import ServicesRepository
-from .components_repository import CardExtensions
-from .actionlog import ActionLog
+from . import view
+from .comp import ActionLog
